Fixed: filter 여러 컴포넌트로 분리
